Rename instanseg file, resolve #52 (#60)
